Slowly getting there.
* Harmonize bukkit's walkspeed with attributes;

* Reset Improbable's data on certain events;

* Check for Improbable on fakefall and fakeground;

* SurvivalFly:
 - Unbound liquid workarounds from velocity jump phase, in the hope that it will fix issues regarding waterfalls
 - Throw in a placeholder estimation for the player movement inputs
 - Make the estimation work, for the most part (at least for just walking/sprinting on ground). Lots of issues still remain
 - vDistAir: switch to minJumpGain as maxJumpGain's motion is way too high with jump amplifier
 - Do bunnyfly after the estimation and sharpen bunnyslope
 - On using queued horizontal velocity entries, update Improbable
 - Remove a waterfall workaround (was ineffective)
 - Comments on vDistClimbable

+++ NCPCore/src/main/java/fr/neatmonster/nocheatplus/components/modifier/DummyAttributeAccess.java
@@ -35,7 +35,7 @@ public class DummyAttributeAccess implements IAttributeAccess {
     }
     
     @Override
-    public double getMovementSpeed(final Player player) {
-    	return Double.MAX_VALUE;
+    public float getMovementSpeed(final Player player) {
+    	return Float.MAX_VALUE;
     }
 }
